requirements.txt
flask
transformers
torch
pillow
sentencepiece
torchvision
ultralytics
pillow
numpy